address feedback from the code review
